Adds test in CharactersList component.
